pyproject.toml
[project]
name = "mychatapp"
version = "0.1.0"
description = ""
authors = [
    {name = "manveer11011",email = "manveer11011@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "channels (>=4.3.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "npm (>=0.1.1,<0.2.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
